+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 ### Changed
 
 - Changed hierarchy of classes that communicate with V-REP simulator.
+- Changed arguments for interface to V-REP remote API server.
 - Changed connecting to and disconnecting from V-REP remote API server via
   interface to V-REP remote API server such that establishing multiple
   connections is explicitly forbidden.